requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-Smorest==0.42.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
passlib==1.7.4
marshmallow==3.20.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2==2.9.9
